move socket control to separate file, define game settings, define game state object

+++ server/Controllers/SessionController.ts
@@ -13,15 +13,15 @@ const SessionController = {
       newSession.userId = hash;
       newSession.save();
       res.locals.sessionId = hash;
-      next();
+      return next();
     } catch (error) {
-      next({errMsg: error});
+      return next({errMsg: error});
     }
   },
   //check mongoDB for room creation availability for session
   verifyRoomAvail: async (req: Request, res: Response, next: NextFunction) => {
     try {
-      const sessionId = req.body.stuff.slice(10);
+      const sessionId = req.cookies.sessionId;
       const returnedTest = await sessionInfo.find({userId: sessionId})
       if(returnedTest.length>0){
         if(returnedTest[0].roomStatus === false){
@@ -32,19 +32,22 @@ const SessionController = {
           res.locals.check = {roomAvailable: false}
         }
       }
-      next();
+      console.log(req.body.stuff)
+      console.log('ROOM CHECK')
+      console.log('==================')
+      console.log(res.locals.check)
+      console.log('==================')
+      return next();
     } catch (error) {
-      next({errMsg: error});
+      return next({errMsg: error});
     }
   },
   //check if they have a socket already
   verifySocketAvail: async (req: Request, res: Response, next: NextFunction) => {
     try {
-      const sessionId = req.body.stuff.slice(10);
+      const sessionId = req.cookies.sessionId;
       const returnedTest = await sessionInfo.find({userId: sessionId})
-      console.log(returnedTest)
       if(returnedTest.length>0){
-        console.log('this one '+ returnedTest[0].socketStatus)
         if(returnedTest[0].socketStatus === false){
           await sessionInfo.findOneAndUpdate({userId: sessionId}, {socketStatus: true})
           res.locals.check.socketStatus = true;
@@ -53,26 +56,40 @@ const SessionController = {
           res.locals.check.socketStatus = false;
         }
       }
-      next();
+      console.log('SOCKET CHECK')
+      console.log('==================')
+      console.log(res.locals.check)
+      console.log('==================')
+
+      return next();
     } catch (error) {
       const sessionId = req.body.stuff.slice(10);
       await sessionInfo.findOneAndUpdate({userId: sessionId}, {roomStatus: false})
-      next({errMsg: error});
+      return next({errMsg: error});
     }
   },
   //close session by deleting from mongodb
   closeSession: async (req: Request, res: Response, next: NextFunction) => {
     try {
       //get sessionId from request body
-      const sessionId = req.body.stuff.slice(10);
+      const sessionId = req.cookies.sessionId;
+      if(sessionId === ''){
+        return next();
+      }
       console.log('Closing session: ' + sessionId);
       await sessionInfo.findOneAndDelete(sessionId)
-      .then((resp) => {
-        console.log('deleted')
-      });
-      next();
+      return next();
     } catch (error) {
-      next({errMsg: error})
+      return next({errMsg: error})
+    }
+  },
+  resetSocket: async (req: Request, res: Response, next: NextFunction) => {
+    try {
+      const sessionId = req.cookies.sessionId;
+      await sessionInfo.findOneAndUpdate({userId: sessionId}, {socketStatus: false, roomStatus: false});
+      return next()
+    } catch (error) {
+      return next({errMsg: error})
     }
   }
 }
